Some code restructoring
