+++ frontend/src/app/directives/sortable.directive.ts
@@ -1,0 +1,31 @@
+import {Directive, EventEmitter, Input, Output} from '@angular/core';
+import { Alarm } from 'src/app/models/alarm';
+
+export type SortColumn = keyof Alarm | '';
+export type SortDirection = 'asc' | 'desc' | '';
+const rotate: {[key: string]: SortDirection} = { 'asc': 'desc', 'desc': '', '': 'asc' };
+
+export interface SortEvent {
+    column: SortColumn;
+    direction: SortDirection;
+}
+
+@Directive({
+    selector: 'th[sortable]',
+    host: {
+        '[class.asc]': 'direction === "asc"',
+        '[class.desc]': 'direction === "desc"',
+        '(click)': 'rotate()'
+    }
+})
+export class SortableDirective {
+
+    @Input() sortable: SortColumn = '';
+    @Input() direction: SortDirection = '';
+    @Output() sort = new EventEmitter<SortEvent>();
+
+    rotate(): void {
+        this.direction = rotate[this.direction];
+        this.sort.emit({column: this.sortable, direction: this.direction});
+    }
+}
